Added printout for sys_execve - buggy
Added printout for execve system calls in jsys_execve(). However, whenever this function is called the system crashes. This is odd, because the code worked on the first couple of runs, and now it consistently crashes. Perhaps the VM I am using has been corrupted a bit too much and can no longer handle the module?