Fixed printout bug for sys_execve
In the previous commit, there was a bug in for printing the filename field of a pH_profile. The problem was the filename parameter for sys_execve is not terminated by the \0 character, which leads to undefined behaviour when the string is passed to printk(). This bug was fixed by using a char array of \0 characters, and filling the array with all sanitary characters in filename.