Add files via upload
Added files for initial attempt at porting pH to running on a real machine instead of a VM. However, at the time of this commit, there are know bugs. The module currently often crashes the system. I assume this is because of either some erroneous code in the module itself, or that the issue is caused by one of the system calls included in system_call_prototypes.h.