Update km.c - this version is close to working
This version is quite close to working. It was tested on the development box (wells), and worked fine for a while, but then I was able to make the system crash. Since then I have also been able to crash it on the testing VMs, although it is a bit difficult to do so. Errors I have encountered include int3 and page faults. For further information regarding crashes, refer to /home/shane/Documents/crash_log_for_pH_port3.